feat: working bronze ingestion
Note that this separates seeding outside of dagster, as I thought that the DLT failure was due to the Dagster / DLT integration.  However, it turns out it was due to library incompatibilities in destinations.  Cleaning up the venv and corrupted duckdb file fixed this.  Will try once more to integrate back-to-back DLT flows in dagster as final ingestion pipeline.